// src/api/stetsom/model/completeUploadInput.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { CompleteUploadInputAlt } from "./completeUploadInputAlt";
import type { CompleteUploadInputType } from "./completeUploadInputType";

export interface CompleteUploadInput {
  /**
   * @minLength 1
   * @maxLength 255
   */
  filename: string;
  /** @maxLength 2048 */
  file_url: string;
  type: CompleteUploadInputType;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  size_bytes: number;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  width?: number;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  height?: number;
  alt?: CompleteUploadInputAlt;
  tags?: string[];
}